doWhat bug tweaked
The doWhat function has been tweaked to be more flexible. The previous solution assumes there will ever only be two comma separated items in the text file and that those two items are related. However, this might not be the case if more items are added and some items which do not necessarily require other items to run. The helper function has been tweaked to add instructions for adding items to the text file.